Elfangor567's work (#202)
* Fake PR

I apologize for the bad coding.

* Phase 2

Still not done, it needs to be outputted, I'll work on it when I have time.

* Update

Not even close to done, but an update as to what is happening in the code.

* Phase 3

There are as many phases as there. I know how many there are. It doesn't matter.

* Loops and Sets

Cuts down nicely on what would have been a huge list of if statements.

* Phase 4

I think I have most of it down, however, I need to get certain strings from a place other than the .ck2 file.

* Little More

* Phase 5

Now I can build it. Shame it doesn't work yet.

* Removed Outermost Block

* Phase 6

I'm Elfangor567 and this is my least favorite phase on the Citadel.

* Phase 7

I don't think the engines can take anymore cap'n!

* Phase 8

We're almost out.

* Phase 9

IT'S ALIVE

* Testing Problems

- 3 errors related to testing
- 2 deal with the .obj created by the two new .cpp files

* Phase 10

RISE, MY CREATION!

* Tweaks

- Travis, you better like these changes.

* &

I'll miss your reference...

* Everything But Flags

Don't worry, the flags will be coming to plant themselves in your soil very soon.

* Flag Class

Why is THIS the hard part!?

* Flags is Now a Class

I hope this fills your brain with happy chemicals.

* Tweaks

Tweaked

* More Tweaks

* Colour Problems

Black is not the new anything.

* Colour Finally FIxed

Huzzah

* Common

* Removed Debug

Almost forgot about this

* Update commons

Co-authored-by: Elfangor567 <gliamh@hotmail.com>
Co-authored-by: Stephen May <idhrendur@gmail.com>